Better handle API failures
The previous implementation just let the error bubble up. This means
we might loose a sale due to sales tax. Most stores would rather just
pay the sales tax out-of-pocket in these rare events over loosing
a sale.

It might be worth making this an option if there are stores that prefer
to loose the sale but I'll leave it to those stores to implement that.

The system basically treats it as if the order it outside the nexus
(no tax collected) if it cannot contact the API.

To ensure accurate reporting it leaves a log message that is prefixed
with "Taxjar Failure". THe logs should be searched before reporting
sales tax to see if there are any of these.

If a more active notification is desired a log service can be used to
look for that string and send an e-mail to the right person when it
happens.